requirements.txt
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
Django==2.0
django-debug-toolbar==1.11
django-shortuuidfield==0.1.3
djangorestframework==3.9.1
freeze==1.0.10
idna==2.8
mysqlclient==1.4.1
Pillow==5.4.1
python-memcached==1.59
pytz==2018.9
qiniu==7.2.2
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.24.1
webencodings==0.5.1
